Proper inheritance level for writing-mode, Fixes #98
