+++ org/w3c/css/properties/CSS3SVGProperties.properties
@@ -19,7 +19,8 @@ marker-end:                             org.w3c.css.properties.svg.CssMarkerEnd
 marker-mid:                             org.w3c.css.properties.svg.CssMarkerMid
 marker-start:                           org.w3c.css.properties.svg.CssMarkerStart
 kerning:                                org.w3c.css.properties.svg.CssKerning
-writing-mode:                           org.w3c.css.properties.svg.CssWritingMode
+# replaced with css3
+#writing-mode:                           org.w3c.css.properties.svg.CssWritingMode
 clip-path:                              org.w3c.css.properties.svg.CssClipPath
 clip-rule:                              org.w3c.css.properties.svg.CssClipRule
 mask:                                   org.w3c.css.properties.svg.CssMask
@@ -378,6 +379,7 @@ ruby-position:                          org.w3c.css.properties.css3.CssRubyPosit
 ruby-align:                             org.w3c.css.properties.css3.CssRubyAlign
 ruby-overhang:                          org.w3c.css.properties.css3.CssRubyOverhang
 alignment-adjust:                       org.w3c.css.properties.css3.CssAlignmentAdjust
+writing-mode:                           org.w3c.css.properties.css3.CssWritingMode
 fit:                                    org.w3c.css.properties.css3.CssFit
 handheld.background-clip:               org.w3c.css.properties.css3.CssBackgroundClip
